requirements.txt
requests>=2.31.0
msal>=1.24.0
pyyaml>=6.0
rich>=13.0.0
jinja2>=3.1.0
